[#104] fix - 상세페이지 문제 레벨 수정 및 코드 정렬
