+++ OrrRock/OrrRock/ViewControllers/VideoCollectionViewController/VideoCollectionView+VideoCollectionViewExtensions.swift
@@ -115,11 +115,7 @@ extension VideoCollectionViewController  : UICollectionViewDataSource{
             cell.cellLabel.backgroundColor = .orrFail
         }
         
-        if videoInformationArray[indexPath.item].problemLevel == -1{
-            cell.cellLabel.text = "V?"
-        } else {
-            cell.cellLabel.text = "V\(videoInformationArray[indexPath.item].problemLevel)"
-        }
+        cell.cellLabel.text = videoInformationArray[indexPath.item].problemLevel == -1 ? "V?" : "V\(videoInformationArray[indexPath.item].problemLevel)"
 
         return cell
     }
